lisence_plate_detection_built_on_streamlit_images/requirements.txt
streamlit
opencv-python-headless
opencv-python
numpy 
Pillow
ultralytics
paddleocr
paddlepaddle
